refactor: move graph model to toplevel
