+++ src/cbrkit/model/__init__.py
@@ -1,6 +1,9 @@
+from . import graph
 from .result import CycleResult, QueryResultStep, Result, ResultStep
 
 __all__ = [
+    "result",
+    "graph",
     "QueryResultStep",
     "ResultStep",
     "Result",
