+++ src/cbrkit/sim/graphs/__init__.py
@@ -1,48 +1,18 @@
-from ...helpers import optional_dependencies
 from . import astar
+from .alignment import dtw, smith_waterman
 from .brute_force import brute_force
+from .common import ElementMatcher, GraphSim, default_element_matcher
 from .isomorphism import isomorphism
-from .model import (
-    Edge,
-    Graph,
-    GraphSim,
-    Node,
-    SerializedEdge,
-    SerializedGraph,
-    SerializedNode,
-    dump,
-    from_dict,
-    is_sequential,
-    load,
-    to_dict,
-)
 from .precompute import precompute
-
-with optional_dependencies():
-    from .model import from_rustworkx, to_rustworkx
-
-with optional_dependencies():
-    from .model import from_networkx, to_networkx
 
 __all__ = [
     "astar",
     "brute_force",
     "isomorphism",
     "precompute",
-    "Node",
-    "Edge",
-    "Graph",
+    "dtw",
+    "smith_waterman",
     "GraphSim",
-    "SerializedNode",
-    "SerializedEdge",
-    "SerializedGraph",
-    "load",
-    "dump",
-    "to_dict",
-    "from_dict",
-    "is_sequential",
-    "to_rustworkx",
-    "from_rustworkx",
-    "to_networkx",
-    "from_networkx",
+    "ElementMatcher",
+    "default_element_matcher",
 ]
